Final Project For AP Computer Science. It only works on BlueJ I think
This Code is a super simple version of pokemon. Made for AP Computer
Science.

ONLY WORKS WITH BLUEJ

PLEASE DONT DOWNLOAD THIS CODE FOR A CSE CLASS AND USE IT AS YOURS...
DOWNLOADING THIS CODE AND USING IT AS YOURS WILL NOT HELP ANY
STUDENT...ESPECIALLY NOT CSE MAJORS
THIS CODE IS FOR HELP AND I KINDA DIDN'T WANT IT ON MY COMPUTER SO I
THREW IT UP HERE LOL